Extract shared barista module and reorganize examples
Move webrtc/, chaos/, e2e/ under examples/pipecat/barista/ with a single
pyproject.toml. Extract duplicated code (tools schema, system prompt, tool
handlers, service creation, pipeline builder) into shared.py.
