+++ examples/pipecat/barista/e2e/README.md
