+++ examples/pipecat/barista/e2e/customer.py
@@ -1,4 +1,7 @@
 import os
+import sys
+
+sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
 from loguru import logger
 from pipecat.audio.vad.silero import SileroVADAnalyzer
