+++ examples/pipecat/barista/env.example
