--- examples/pipecat/chaos/bot.py
@@ -1,287 +1,0 @@
-#
-# Copyright (c) 2024–2025, Daily
-#
-# SPDX-License-Identifier: BSD 2-Clause License
-#
-
-import os
-import random
-from typing import AsyncGenerator
-
-from dotenv import load_dotenv
-from loguru import logger
-from pipecat.adapters.schemas.function_schema import FunctionSchema
-from pipecat.adapters.schemas.tools_schema import ToolsSchema
-from pipecat.audio.vad.silero import SileroVADAnalyzer
-from pipecat.frames.frames import Frame, LLMRunFrame, TTSSpeakFrame
-from pipecat.pipeline.pipeline import Pipeline
-from pipecat.pipeline.runner import PipelineRunner
-from pipecat.pipeline.task import PipelineParams, PipelineTask
-from pipecat.processors.aggregators.llm_context import LLMContext
-from pipecat.processors.aggregators.llm_response_universal import (
-    LLMContextAggregatorPair,
-)
-from pipecat.runner.types import RunnerArguments
-from pipecat.runner.utils import create_transport
-from pipecat.services.cartesia.tts import CartesiaTTSService
-from pipecat.services.deepgram.stt import DeepgramSTTService
-from pipecat.services.llm_service import FunctionCallParams
-from pipecat.services.openai.llm import OpenAILLMService
-from pipecat.transports.base_transport import BaseTransport, TransportParams
-from pipecat.transports.daily.transport import DailyParams
-from pipecat.transports.websocket.fastapi import FastAPIWebsocketParams
-
-import finchvox
-from finchvox import FinchvoxProcessor
-
-load_dotenv(override=True)
-
-finchvox.init(service_name="pipecat-chaos-demo")
-
-
-def should_fail(rate_env_var: str) -> bool:
-    rate = float(os.getenv(rate_env_var, "0"))
-    if rate > 0 and random.random() < rate:
-        logger.warning(f"[CHAOS] Injecting failure ({rate_env_var}={rate})")
-        return True
-    return False
-
-
-def with_chaos(handler):
-    async def wrapped(params: FunctionCallParams):
-        if should_fail("TOOL_FAIL_RATE"):
-            await params.result_callback(
-                {"error": "Simulated tool failure", "success": False}
-            )
-            return
-        await handler(params)
-
-    return wrapped
-
-
-class ChaosOpenAILLMService(OpenAILLMService):
-    async def get_chat_completions(self, params):
-        if should_fail("LLM_FAIL_RATE"):
-            raise Exception("Simulated LLM failure")
-        return await super().get_chat_completions(params)
-
-
-class ChaosDeepgramSTTService(DeepgramSTTService):
-    async def run_stt(self, audio: bytes) -> AsyncGenerator[Frame, None]:
-        if should_fail("STT_FAIL_RATE"):
-            raise Exception("Simulated STT failure")
-        async for frame in super().run_stt(audio):
-            yield frame
-
-
-class ChaosCartesiaTTSService(CartesiaTTSService):
-    async def run_tts(self, text: str) -> AsyncGenerator[Frame, None]:
-        if should_fail("TTS_FAIL_RATE"):
-            raise Exception("Simulated TTS failure")
-        async for frame in super().run_tts(text):
-            yield frame
-
-
-async def add_item_to_order(params: FunctionCallParams):
-    item = params.arguments.get("item", "item")
-    size = params.arguments.get("size", "")
-    modifications = params.arguments.get("modifications", [])
-    description = f"{size} {item}".strip()
-    if modifications:
-        description += f" with {', '.join(modifications)}"
-    await params.result_callback({"success": True, "item": description})
-
-
-async def remove_item_from_order(params: FunctionCallParams):
-    item = params.arguments.get("item", "item")
-    await params.result_callback({"success": True, "removed": item})
-
-
-async def get_order_summary(params: FunctionCallParams):
-    await params.result_callback(
-        {"items": ["medium oat latte", "blueberry muffin"], "item_count": 2}
-    )
-
-
-async def submit_order(params: FunctionCallParams):
-    name = params.arguments.get("customer_name", "friend")
-    await params.result_callback({"success": True, "order_number": 47, "name": name})
-
-
-transport_params = {
-    "daily": lambda: DailyParams(
-        audio_in_enabled=True,
-        audio_out_enabled=True,
-        vad_analyzer=SileroVADAnalyzer(),
-    ),
-    "twilio": lambda: FastAPIWebsocketParams(
-        audio_in_enabled=True,
-        audio_out_enabled=True,
-        vad_analyzer=SileroVADAnalyzer(),
-    ),
-    "webrtc": lambda: TransportParams(
-        audio_in_enabled=True,
-        audio_out_enabled=True,
-        vad_analyzer=SileroVADAnalyzer(),
-    ),
-}
-
-
-async def run_bot(transport: BaseTransport):
-    stt = ChaosDeepgramSTTService(api_key=os.getenv("DEEPGRAM_API_KEY"))
-
-    tts = ChaosCartesiaTTSService(
-        api_key=os.getenv("CARTESIA_API_KEY"),
-        voice_id="f786b574-daa5-4673-aa0c-cbe3e8534c02",
-    )
-
-    llm = ChaosOpenAILLMService(
-        api_key=os.getenv("OPENAI_API_KEY"),
-        params=ChaosOpenAILLMService.InputParams(temperature=0.5),
-    )
-
-    llm.register_function("add_item_to_order", with_chaos(add_item_to_order))
-    llm.register_function("remove_item_from_order", with_chaos(remove_item_from_order))
-    llm.register_function("get_order_summary", with_chaos(get_order_summary))
-    llm.register_function("submit_order", with_chaos(submit_order))
-
-    @llm.event_handler("on_function_calls_started")
-    async def on_function_calls_started(service, function_calls):
-        await tts.queue_frame(TTSSpeakFrame("One sec."))
-
-    add_item_schema = FunctionSchema(
-        name="add_item_to_order",
-        description="Add an item to the customer's order",
-        properties={
-            "item": {
-                "type": "string",
-                "description": "The item being ordered (e.g., latte, cappuccino, blueberry muffin)",
-            },
-            "size": {
-                "type": "string",
-                "enum": ["small", "medium", "large"],
-                "description": "Size of the drink if applicable",
-            },
-            "modifications": {
-                "type": "array",
-                "items": {"type": "string"},
-                "description": "Modifications like 'oat milk', 'extra shot', 'no foam'",
-            },
-        },
-        required=["item"],
-    )
-
-    remove_item_schema = FunctionSchema(
-        name="remove_item_from_order",
-        description="Remove an item from the customer's order",
-        properties={
-            "item": {
-                "type": "string",
-                "description": "The item to remove",
-            },
-        },
-        required=["item"],
-    )
-
-    get_order_summary_schema = FunctionSchema(
-        name="get_order_summary",
-        description="Get a summary of what's currently in the customer's order",
-        properties={},
-        required=[],
-    )
-
-    submit_order_schema = FunctionSchema(
-        name="submit_order",
-        description="Submit the order for preparation. Call this when the customer is done ordering.",
-        properties={
-            "customer_name": {
-                "type": "string",
-                "description": "The customer's name for the order",
-            },
-        },
-        required=["customer_name"],
-    )
-
-    tools = ToolsSchema(
-        standard_tools=[
-            add_item_schema,
-            remove_item_schema,
-            get_order_summary_schema,
-            submit_order_schema,
-        ]
-    )
-
-    messages = [
-        {
-            "role": "system",
-            "content": """You are Sam, a friendly barista at a cozy neighborhood coffee shop. You're warm and welcoming but efficient - you keep conversations moving without being rushed.
-
-Your menu includes:
-- Coffee drinks: espresso, lattes, cappuccinos, americanos, mochas, cold brew, drip coffee
-- Tea: green, black, chai, herbal (hot or iced)
-- Pastries: muffins, croissants, scones, cookies
-- Food: bagels, breakfast sandwiches, turkey or veggie sandwiches
-
-Behavior:
-- Greet customers warmly and ask what they'd like
-- Use the order tools to manage their order - don't just pretend
-- Ask clarifying questions when needed (size, hot/iced, milk preference)
-- Only suggest options or describe items when the customer asks - no unsolicited upselling
-- Keep responses short and natural since this is a voice conversation
-- When the customer is done, ask for their name and submit the order
-
-Your output will be converted to audio so don't include special characters. Be conversational and brief.""",
-        },
-    ]
-
-    context = LLMContext(messages, tools)
-    context_aggregator = LLMContextAggregatorPair(context)
-
-    pipeline = Pipeline(
-        [
-            transport.input(),
-            stt,
-            context_aggregator.user(),
-            llm,
-            tts,
-            transport.output(),
-            FinchvoxProcessor(),
-            context_aggregator.assistant(),
-        ]
-    )
-
-    task = PipelineTask(
-        pipeline,
-        params=PipelineParams(
-            enable_metrics=True,
-            enable_usage_metrics=True,
-        ),
-        enable_tracing=True,
-        enable_turn_tracking=True,
-    )
-
-    @transport.event_handler("on_client_connected")
-    async def on_client_connected(transport, client):
-        logger.info("Client connected")
-        await task.queue_frames([LLMRunFrame()])
-
-    @transport.event_handler("on_client_disconnected")
-    async def on_client_disconnected(transport, client):
-        logger.info("Client disconnected")
-        await task.cancel()
-
-    runner = PipelineRunner(handle_sigint=False)
-
-    await runner.run(task)
-
-
-async def bot(runner_args: RunnerArguments):
-    """Main bot entry point compatible with Pipecat Cloud."""
-    transport = await create_transport(runner_args, transport_params)
-    await run_bot(transport)
-
-
-if __name__ == "__main__":
-    from pipecat.runner.run import main
-
-    main()
