--- examples/pipecat/chaos/env.example
@@ -1,18 +1,0 @@
-# Chaos/Failure injection rates (0.0-1.0)
-# Set to 0 or leave unset for normal operation
-LLM_FAIL_RATE=0
-STT_FAIL_RATE=0
-TTS_FAIL_RATE=0
-TOOL_FAIL_RATE=0
-
-# Required API keys (copy from parent example)
-DEEPGRAM_API_KEY=your_deepgram_key
-CARTESIA_API_KEY=your_cartesia_key
-OPENAI_API_KEY=your_openai_key
-
-# Set to any value to enable tracing
-ENABLE_TRACING=true
-# OTLP endpoint (defaults to localhost:4317 if not set)
-OTEL_EXPORTER_OTLP_ENDPOINT=http://localhost:4317
-# Set to any value to enable console output for debugging
-# OTEL_CONSOLE_EXPORT=true
