--- examples/pipecat/chaos/pyproject.toml
@@ -1,13 +1,0 @@
-[project]
-name = "pipecat-finchvox-chaos"
-version = "0.0.1"
-description = "A Pipecat example with chaos/failure injection for testing"
-requires-python = ">=3.11,<3.14"
-dependencies = [
-    "pipecat-ai[daily,webrtc,websocket,silero,cartesia,deepgram,openai,tracing,runner]",
-    "finchvox",
-]
-
-[tool.uv.sources]
-finchvox = { path = "../../..", editable = true }
-pipecat-ai = { path = "/Users/derek/projects/pipecat", editable = true }
