--- examples/pipecat/webrtc/env.example
@@ -1,10 +1,0 @@
-DEEPGRAM_API_KEY=your_deepgram_key
-CARTESIA_API_KEY=your_cartesia_key
-OPENAI_API_KEY=your_openai_key
-
-# Set to any value to enable tracing
-ENABLE_TRACING=true
-# OTLP endpoint (defaults to localhost:4317 if not set)
-OTEL_EXPORTER_OTLP_ENDPOINT=http://localhost:4317
-# Set to any value to enable console output for debugging
-# OTEL_CONSOLE_EXPORT=true
